Lots of Changes
The cps reading moves depending on the suffix so  more work must be done on stabilizing he result.

Shop Readings in progress, should search the area of the shop item for the little icon then read the number close to it. This must be done so if buying 10x the cost can still be found